minutes: split out BoD logic from general meetings logic
