+++ staff/meetings/meetings.py
@@ -1,18 +1,9 @@
-import getpass
 import os
 import os.path
 import re
-import shutil
 import time
-from collections import defaultdict
 from datetime import datetime
-from math import ceil
 from os.path import join
-from string import Template
-
-
-# Max length of usernames
-USERNAME_LEN = 16
 
 
 def get_minutes_folder():
@@ -24,27 +15,9 @@ def get_minutes_folder():
     return os.path.expanduser('~staff/minutes')
 
 
-def ls(state='all'):
-    """Lists everyone with the given membership state.
-
-    Args:
-        state: the membership state. 'all' lists all attendees regardless of
-               state.
-
-    Returns:
-        A list of BoD attendees with the given membership state
-
-    """
-    states = get_bod_membership()
-    if state != 'all':
-        states = {k: v for k, v in states.items() if v == state}
-    return states
-
-
-def quorum():
-    """Returns the quorum for this BoD meeting."""
-    bod = ls(state='bod')
-    return int(ceil(2 / 3 * len(bod)))
+def get_minutes_choices():
+    """Returns the choices available for meeting type."""
+    return os.listdir(get_minutes_folder())
 
 
 def get_template(choice):
@@ -53,11 +26,6 @@ def get_template(choice):
     if os.path.exists(join(minutes_folder, choice, 'template')):
         return join(minutes_folder, choice, 'template')
     return join(minutes_folder, 'template')
-
-
-def get_minutes_choices():
-    """Returns the choices available for meeting type."""
-    return os.listdir(get_minutes_folder())
 
 
 def get_semester():
@@ -82,9 +50,19 @@ def get_semester():
     return join(str(year), sem)
 
 
-def get_bod_minutes_path(semester=get_semester()):
-    """Gets the path to the BoD minutes directory for the given semester."""
-    return get_minutes_path('bod', semester=semester)
+def get_prev_semester(semester=get_semester()):
+    """Returns the path for the semester before the given semester.
+
+    >>> get_prev_semester('2017/Fall')
+    '2017/Spring'
+    >>> get_prev_semester('2017/Spring')
+    '2016/Fall'
+    """
+    year, sem = semester.split('/')
+    if sem == 'Spring':
+        return join(str(int(year) - 1), 'Fall')
+    else:
+        return join(year, 'Spring')
 
 
 def get_minutes_path(choice, semester=get_semester()):
@@ -106,33 +84,6 @@ def get_minutes_path(choice, semester=get_semester()):
 def get_minutes_file():
     """Gets the filename for the current minutes."""
     return time.strftime('%Y-%m-%d')
-
-
-def get_bod_membership_file(semester=get_semester()):
-    """Gets the BoD membership file for the given semester."""
-    folder = get_bod_minutes_path(semester=semester)
-    path = join(folder, 'membership')
-    return path
-
-
-def get_bod_membership(semester=get_semester()):
-    """Get the membership status for every BoD attendee for the given semester.
-
-    Args:
-        semester: The specified semester. Defaults to the current semester.
-
-    Returns:
-        A dictionary of (attendee, state) mappings.
-    """
-    status = defaultdict(int)
-    path = get_bod_membership_file(semester=semester)
-    if not os.path.exists(path):
-        new_semester()
-    with open(path) as f:
-        for line in f:
-            user, state = line.split()
-            status[user] = state
-    return status
 
 
 def get_minutes(folder):
@@ -171,127 +122,3 @@ def get_attendance(path):
             attended.add(lines[i])
             i += 1
     return attended
-
-
-def get_prev_semester(semester=get_semester()):
-    """Returns the path for the semester before the given semester.
-
-    >>> get_prev_semester('2017/Fall')
-    '2017/Spring'
-    >>> get_prev_semester('2017/Spring')
-    '2016/Fall'
-    """
-    year, sem = semester.split('/')
-    if sem == 'Spring':
-        return join(str(int(year) - 1), 'Fall')
-    else:
-        return join(year, 'Spring')
-
-
-def new_semester():
-    """Sets up a semester's directory.
-
-    Currently copies the BoD membership file from the previous semester.
-    """
-    new_file = get_bod_membership_file()
-    old_file = get_bod_membership_file(semester=get_prev_semester())
-    shutil.copyfile(old_file, new_file)
-    os.chmod(new_file, 0o664)
-
-
-def minutes_setup(notes, choice):
-    # Create minutes file from template if it doesn't exist
-    if not os.path.exists(notes):
-        shutil.copyfile(get_template(choice), notes)
-        os.chmod(notes, 0o644)
-
-    with open(notes, 'r') as f:
-        s = Template(f.read())
-
-    # Substitute values
-    subs = {'username': getpass.getuser(), 'start_time': time.strftime('%H:%M')}
-    if choice == 'bod':
-        subs['quorum'] = str(quorum())
-    s = s.safe_substitute(subs)
-
-    # Write out file with substituted values
-    with open(notes, 'w') as f:
-        f.write(s)
-
-
-def minutes_done(notes, choice):
-    # Substitute in ending values
-    with open(notes, 'r') as f:
-        s = Template(f.read())
-    s = s.safe_substitute(end_time=time.strftime('%H:%M'))
-
-    # Write out file with substituted values
-    with open(notes, 'w') as f:
-        f.write(s)
-
-    if choice == 'bod':
-        update_membership()
-
-
-def update_membership():
-    """Updates the membership file for the current semester.
-
-    Uses the attendance list for the current meeting to update the membership
-    file for the current semester.
-    """
-    cur_status = get_bod_membership()
-    bod_members = {member for member, status in cur_status.items()
-                   if status == 'bod'}
-
-    minutes_path = get_bod_minutes_path()
-    minutes_files = get_minutes(minutes_path)
-    if minutes_files:
-        print(join(minutes_path, minutes_files[-1]))
-        attendees_today = get_attendance(join(minutes_path, minutes_files[-1]))
-
-    # Start new status dict. Begin by adding all BoD members.
-    new_status = {member: status for member, status in cur_status.items()
-                  if status == 'bod'}
-
-    # Add new members to BoD
-    guests = attendees_today - bod_members
-    for attendee in guests:
-        reply = input(attendee + ' is not on BoD. Would they like to join? '
-                      '(y/n) ')
-        if reply == 'y' or reply == 'yes':
-            new_status[attendee] = 'bod'
-
-    # Kick people off BoD if they haven't come to the previous 2 meetings
-    # (including the current one)
-    if len(minutes_files) >= 2:
-        print(join(minutes_path, minutes_files[-2]))
-        attended_last_meeting = get_attendance(join(minutes_path,
-                                                    minutes_files[-2]))
-    else:
-        last_semester_minutes_path = get_bod_minutes_path(get_prev_semester())
-        last_semester_minutes = get_minutes(last_semester_minutes_path)
-        assert len(last_semester_minutes) > 1, \
-            'Expecting at least one BoD meeting last semester'
-        attended_last_meeting = get_attendance(join(last_semester_minutes_path,
-                                                    last_semester_minutes[-1]))
-
-    print(attendees_today)
-    print(attended_last_meeting)
-    for non_attendee in bod_members - (attendees_today | attended_last_meeting):
-        print(non_attendee, 'is being kicked off bod')
-        del new_status[non_attendee]
-
-    # Write out new status
-    def membership_key(p):
-        special = {'bod': 0, 'offbod': 1}
-        u, s = p
-        if s in special:
-            s = special[s]
-        return (s, u)
-    sorted_new_status = sorted(new_status.items(), key=membership_key)
-    str_new_status = ''.join('{}  {}\n'.format(attendee.ljust(USERNAME_LEN),
-                                               status)
-                             for attendee, status in sorted_new_status)
-
-    with open(get_bod_membership_file(), 'w') as f:
-        f.write(str_new_status)
